agent-examples/04-llamaindex-query-agent/requirements.txt
python-dotenv
llama-index
